rename header to logging.h
