+++ stitcher/include/airmap/logging.h
